refactor: refactor parsers
affects: @takomo/stacks-config
