+++ packages/stacks-config/src/index.ts
@@ -1,1 +1,2 @@
-export * from "./parser"
+export { parseStackConfigFile } from "./parse-stack-config-file"
+export { parseStackGroupConfigFile } from "./parse-stack-group-config-file"
