+++ packages/stacks-config/src/parse-depends.ts
@@ -1,0 +1,9 @@
+import { StackPath } from "@takomo/core"
+
+export const parseDepends = (value: any): StackPath[] => {
+  if (value === null || value === undefined) {
+    return []
+  }
+
+  return Array.isArray(value) ? value : [value]
+}
