+++ packages/stacks-config/src/parse-ignore.ts
@@ -1,0 +1,7 @@
+export const parseIgnore = (value: unknown): boolean | null => {
+  if (value === null || value === undefined) {
+    return null
+  }
+
+  return value === true
+}
