+++ packages/stacks-config/src/parse-template-bucket.ts
@@ -1,0 +1,16 @@
+import { TemplateBucketConfig } from "@takomo/stacks-model"
+
+export const parseTemplateBucket = (
+  value: any,
+): TemplateBucketConfig | null => {
+  if (!value) {
+    return null
+  }
+
+  const { name, keyPrefix } = value
+
+  return {
+    name: name || null,
+    keyPrefix: keyPrefix || null,
+  }
+}
