+++ packages/stacks-config/src/parse-termination-protection.ts
@@ -1,0 +1,7 @@
+export const parseTerminationProtection = (value: unknown): boolean | null => {
+  if (value === null || value === undefined) {
+    return null
+  }
+
+  return value === true
+}
