+++ packages/stacks-config/test/stack-group-config-1.yml
